+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `libc` functions:
-  - Memory functions: `memcpy`.
+  - Memory functions: `memcpy`, `memmove`.
 
 ## [v0.1.0] - 2023-11-07